+++ Source/Assemblies/Http/HttpClientInvoker/Roblox.Http.HttpClientInvoker/Models/IClientRequest.ts
@@ -1,4 +1,4 @@
-import { HttpRequestMethodEnum } from '../Enumeration/HttpRequestMethodEnum';
+import { HttpRequestMethodEnum } from '../../../Roblox.Http/Enumeration/HttpRequestMethodEnum';
 import { IHttpHeaders } from './JSON/IHttpHeaders';
 import { IHttpQueries } from './JSON/IHttpQueries';
