+++ Source/Assemblies/Http/HttpClientInvoker/Roblox.Http.HttpClientInvoker/Models/JSON/IHttpQueries.ts
