+++ Source/Assemblies/Http/Roblox.Http/Enumeration/HttpRequestMethodEnum.ts
